+++ projects/ngx-darkbox-gallery/src/lib/config/image-configuration.default.ts
@@ -1,5 +1,0 @@
-import { ImageConfiguration } from '../model/image-configuration';
-
-export class DefaultImageConfiguration implements ImageConfiguration {
-
-}