+++ src/app/app.component.ts
@@ -1,8 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 import { MatIconRegistry } from '@angular/material/icon';
 import { DomSanitizer } from '@angular/platform-browser';
-import { Image } from 'dist/ngx-darkbox-gallery/lib/model/image';
-import { Subject, Subscription } from 'rxjs';
+import { Configuration, Image } from 'projects/ngx-darkbox-gallery/src/public-api';
+import { Subject } from 'rxjs';
 
 @Component({
   selector: 'app-root',
@@ -15,6 +15,8 @@ export class AppComponent implements OnInit {
   eventsSubject: Subject<void> = new Subject<void>();
 
   areMoreImagesAvailable: boolean = true;
+
+  configuration: Configuration;
 
   // Constructor is required here to load the svg images into angular material
   constructor(iconRegistry: MatIconRegistry, sanitizer: DomSanitizer) {