+++ src/app/app.module.ts
@@ -2,17 +2,27 @@ import { HttpClientModule } from '@angular/common/http';
 import { NgModule } from '@angular/core';
 import { FormsModule } from '@angular/forms';
 import { MatButtonModule } from '@angular/material/button';
+import { MatExpansionModule } from '@angular/material/expansion';
 import { MatIconModule } from '@angular/material/icon';
 import { BrowserModule } from '@angular/platform-browser';
 import { NgxDarkboxGalleryModule } from '@failed-successfully/ngx-darkbox-gallery';
 import { AppComponent } from './app.component';
+import { ConfigurationEditorComponent } from './components/configuration-editor/configuration-editor.component';
 import { DarkModeSwitchComponent } from './components/dark-mode-switch/dark-mode-switch.component';
 import { FooterComponent } from './components/footer/footer.component';
+import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
+import { NgxJsonViewerModule } from 'ngx-json-viewer';
+import {MatCheckboxModule} from '@angular/material/checkbox';
+import {MatSelectModule} from '@angular/material/select';
+import {MatInputModule} from '@angular/material/input';
+import { ToastrModule } from 'ngx-toastr';
+
 @NgModule({
   declarations: [
     AppComponent,
     DarkModeSwitchComponent,
-    FooterComponent
+    FooterComponent,
+    ConfigurationEditorComponent
   ],
   imports: [
     BrowserModule,
@@ -21,6 +31,13 @@ import { FooterComponent } from './components/footer/footer.component';
     HttpClientModule,
     MatButtonModule,
     MatIconModule,
+    MatExpansionModule,
+    BrowserAnimationsModule,
+    NgxJsonViewerModule,
+    MatCheckboxModule,
+    MatSelectModule,
+    MatInputModule,
+    ToastrModule.forRoot(),
   ],
   providers: [],
   bootstrap: [AppComponent]