+++ CHANGELOG.md
@@ -8,11 +8,26 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+### Changed
+
+### Removed
+
+
+## [1.0.0-alpha.3]
+
+### Added
+
+* Added configuration editor to demo page (#49, #365)
 * Added autmatic release note extraction to workflows (#409)
 
 ### Changed
 
+* Updated advanced-configuration-example.md (#157)
+
 ### Removed
+
+* Removed unused image-configuration
+
 
 ## [1.0.0-alpha.2] 2022-02-13